Add simulation's template

+++ pkg/cfg.go
@@ -2,8 +2,8 @@ package pkg
 
 type Config struct {
 	Mode    string `yaml:"mode"`
-	Time    uint64 `yaml:"time"`
-	Sources uint64 `yaml:"source"`
-	Devices uint64 `yaml:"device"`
-	Bufsize uint64 `yaml:"bufsize"`
+	Time    int    `yaml:"time"`
+	Sources int    `yaml:"source"`
+	Devices int    `yaml:"device"`
+	Bufsize int    `yaml:"bufsize"`
 }
